move table to subfolder
