+++ resources/views/livewire/delegates/monitor.blade.php
@@ -1,4 +1,4 @@
-<div>
+{{-- <div>
     @if(! count($delegates))
         <div wire:poll="pollDelegates" wire:key="poll_delegates_skeleton">
             <x-tables.desktop.skeleton.delegates.monitor />
@@ -8,4 +8,19 @@
             <x-tables.desktop.delegates.monitor :delegates="$delegates" :round="$round" />
         </div>
     @endif
+</div> --}}
+
+<div
+    id="delegate-monitor-list"
+    class="w-full"
+    wire:init="monitorIsReady"
+    @if ($this->isReady)
+        wire:poll.1s="pollDelegates"
+    @endif
+>
+    <x-skeletons.delegates.monitor>
+        <x-tables.desktop.delegates.monitor />
+
+        <x-tables.mobile.delegates.monitor />
+    </x-skeletons.delegates.monitor>
 </div>
